Add overlay click-to-close functionality and minor adjustments
