+++ src/index.tsx
@@ -1,37 +1,10 @@
 import { createRoot } from 'react-dom/client';
-import { StrictMode, CSSProperties, useState } from 'react';
-import clsx from 'clsx';
-
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
-import { defaultArticleState } from './constants/articleProps';
-
+import { StrictMode } from 'react';
 import './styles/index.scss';
-import styles from './styles/index.module.scss';
+import App from './App';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
-
-const App = () => {
-	const [articleStyles, setArticleStyles] = useState(defaultArticleState);
-
-	return (
-		<main
-			className={clsx(styles.main)}
-			style={
-				{
-					'--font-family': articleStyles.fontFamilyOption.value,
-					'--font-size': articleStyles.fontSizeOption.value,
-					'--font-color': articleStyles.fontColor.value,
-					'--container-width': articleStyles.contentWidth.value,
-					'--bg-color': articleStyles.backgroundColor.value,
-				} as CSSProperties
-			}>
-			<ArticleParamsForm setArticleStyles={setArticleStyles} />
-			<Article />
-		</main>
-	);
-};
 
 root.render(
 	<StrictMode>
